% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{bought_last}
\alias{bought_last}
\title{Find Items Bought in the Last Transaction}
\usage{
bought_last(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the
"E-commerce Data Table Structure" section.}
}
\value{
A data.table of items bought in the most recent transaction.
}
\description{
Find Items Bought in the Last Transaction
}
