% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_average_customer_lifetime}
\alias{calculate_average_customer_lifetime}
\title{Calculate Average Customer Lifetime}
\usage{
calculate_average_customer_lifetime(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}
}
\value{
Average customer lifetime duration.
}
\description{
Calculate Average Customer Lifetime
}
