% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_average_order_value}
\alias{calculate_average_order_value}
\title{Calculate Average Order Value}
\usage{
calculate_average_order_value(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the
"E-commerce Data Table Structure" section.}
}
\value{
Average order value.
}
\description{
Calculate Average Order Value
}
