% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_bulk_items}
\alias{calculate_bulk_items}
\title{Calculate Bulk Items Sold}
\usage{
calculate_bulk_items(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the
"E-commerce Data Table Structure" section.}
}
\value{
A data.table with total quantity and descriptions, sorted in descending order.
}
\description{
Calculate Bulk Items Sold
}
