% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_no_of_invoices_per_month}
\alias{calculate_no_of_invoices_per_month}
\title{Calculate Number of Invoices Per Month}
\usage{
calculate_no_of_invoices_per_month(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}
}
\value{
A data.table with the number of invoices and sales percentage per month.
}
\description{
Calculate Number of Invoices Per Month
}
