% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_return_customers}
\alias{calculate_return_customers}
\title{Calculate Returning Customers}
\usage{
calculate_return_customers(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}
}
\value{
A data.table with the number of invoices per customer.
}
\description{
Calculate Returning Customers
}
