% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_revenue_for}
\alias{calculate_revenue_for}
\title{Calculate Revenue for a Specific Column Value}
\usage{
calculate_revenue_for(x, col, val)
}
\arguments{
\item{x}{A data.table following the structure defined in the
"E-commerce Data Table Structure" section.}

\item{col}{The column name to filter on.}

\item{val}{The value to filter for.}
}
\value{
Total revenue for the specified column value.
}
\description{
Calculate Revenue for a Specific Column Value
}
