% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_revenue_for_customer}
\alias{calculate_revenue_for_customer}
\title{Calculate Revenue for a Customer}
\usage{
calculate_revenue_for_customer(x, selected_customer)
}
\arguments{
\item{x}{A data.table following the structure defined in the
"E-commerce Data Table Structure" section.}

\item{selected_customer}{The customer ID.}
}
\value{
Total revenue for the selected customer.
}
\description{
Calculate Revenue for a Customer
}
