% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_revenue_for_product}
\alias{calculate_revenue_for_product}
\title{Calculate Revenue for a Product}
\usage{
calculate_revenue_for_product(x, selected_product)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}

\item{selected_product}{The product stock code.}
}
\value{
Total revenue for the selected product.
}
\description{
Calculate Revenue for a Product
}
