% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_total_revenue}
\alias{calculate_total_revenue}
\title{Calculate Total Revenue}
\usage{
calculate_total_revenue(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}
}
\value{
Total revenue (sum of sales).
}
\description{
Calculate Total Revenue
}
