% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_unique_customers}
\alias{calculate_unique_customers}
\title{Calculate Unique Customers}
\usage{
calculate_unique_customers(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}
}
\value{
Number of unique customers.
}
\description{
Calculate Unique Customers
}
