% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{calculate_unique_products}
\alias{calculate_unique_products}
\title{Calculate Unique Products}
\usage{
calculate_unique_products(x)
}
\arguments{
\item{x}{A data.table following the structure defined in the
"E-commerce Data Table Structure" section.}
}
\value{
Number of unique products.
}
\description{
Calculate Unique Products
}
