% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{get_customer_data}
\alias{get_customer_data}
\title{Get Customer Data}
\usage{
get_customer_data(x, selected_customer)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}

\item{selected_customer}{The customer ID.}
}
\value{
A data.table with data for the selected customer.
}
\description{
Get Customer Data
}
