% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{product_ranking}
\alias{product_ranking}
\title{Product Ranking}
\usage{
product_ranking(x, what = "top", rows = 5)
}
\arguments{
\item{x}{A data.table following the structure defined in the "E-commerce Data Table Structure" section.}

\item{what}{"top" for best sellers, "bottom" for least sold.}

\item{rows}{Number of rows to return.}
}
\value{
A data.table with ranked products.
}
\description{
Product Ranking
}
